+++ packages/react-native-audio-api/common/cpp/core/BaseAudioContext.h
@@ -30,7 +30,7 @@ class BaseAudioContext {
   virtual ~BaseAudioContext() = default;
 
   std::string getState();
-  [[nodiscard]] int getSampleRate() const;
+  [[nodiscard]] float getSampleRate() const;
   [[nodiscard]] double getCurrentTime() const;
   [[nodiscard]] std::size_t getCurrentSampleFrame() const;
   std::shared_ptr<AudioDestinationNode> getDestination();
@@ -41,12 +41,12 @@ class BaseAudioContext {
   std::shared_ptr<BiquadFilterNode> createBiquadFilter();
   std::shared_ptr<AudioBufferSourceNode> createBufferSource();
   static std::shared_ptr<AudioBuffer>
-  createBuffer(int numberOfChannels, int length, int sampleRate);
+  createBuffer(int numberOfChannels, size_t length, float sampleRate);
   std::shared_ptr<PeriodicWave> createPeriodicWave(
       float *real,
       float *imag,
       bool disableNormalization,
-      int length);
+      size_t length);
   std::shared_ptr<AnalyserNode> createAnalyser();
 
   std::shared_ptr<AudioBuffer> decodeAudioDataSource(const std::string &path);
@@ -62,7 +62,7 @@ class BaseAudioContext {
   std::shared_ptr<AudioDecoder> audioDecoder_ {};
 
   // init in AudioContext or OfflineContext constructor
-  int sampleRate_ {};
+  float sampleRate_ {};
   ContextState state_ = ContextState::RUNNING;
   std::shared_ptr<AudioNodeManager> nodeManager_;
